move match candidate expansion methods around
