+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/IndexMaintainerFactory.java
@@ -21,10 +21,13 @@
 package com.apple.foundationdb.record.provider.foundationdb;
 
 import com.apple.foundationdb.annotation.API;
+import com.apple.foundationdb.record.RecordMetaData;
 import com.apple.foundationdb.record.metadata.Index;
 import com.apple.foundationdb.record.metadata.IndexValidator;
+import com.apple.foundationdb.record.query.plan.cascades.MatchCandidate;
 
 import javax.annotation.Nonnull;
+import java.util.Collections;
 
 /**
  * A factory for {@link IndexMaintainer}.
@@ -68,4 +71,45 @@ public interface IndexMaintainerFactory {
      */
     @Nonnull
     IndexMaintainer getIndexMaintainer(@Nonnull IndexMaintainerState state);
+
+    /**
+     * Create {@link MatchCandidate}s to use for the given {@link Index}.
+     * Implementors can assume that the index is one of this factory's
+     * {@linkplain #getIndexTypes() supported types}. Note that this is
+     * structured as a method on the maintainer factory so that indexes
+     * defined outside the core repository can define the structure
+     * used by the planner to match them. However, use cases should be
+     * aware that the planner
+     *
+     * <p>
+     * By default, this returns an empty collection. This means that the
+     * index will not be matchable by the {@link com.apple.foundationdb.record.query.plan.cascades.CascadesPlanner}.
+     * Index types that want to be used in plans should implement this
+     * method. Utility methods in {@link MatchCandidate} can be used to
+     * create, for example, canonical expansions of indexes that behave like
+     * indexes defined in the core repository (e.g., {@link com.apple.foundationdb.record.metadata.IndexTypes#VALUE}
+     * indexes).
+     * </p>
+     *
+     * <p>
+     * Indexes may return more than one {@link MatchCandidate}. This can be
+     * useful for index types that have more than one way of being scanned.
+     * For example, the {@link com.apple.foundationdb.record.metadata.IndexTypes#RANK}
+     * index can be scanned {@link com.apple.foundationdb.record.IndexScanType#BY_VALUE},
+     * at which point it behaves just like a {@link com.apple.foundationdb.record.metadata.IndexTypes#VALUE}
+     * index, or it can be scanned {@link com.apple.foundationdb.record.IndexScanType#BY_RANK},
+     * at which point it behaves very differently. The different scan types are reflected
+     * in different {@link MatchCandidate} which encode different information about
+     * the index and how it should be scanned.
+     * </p>
+     *
+     * @param metaData the meta-data in which the index is defined
+     * @param index the index to expend
+     * @param reverse whether the index is to be scanned in reverse
+     * @return a collection of {@link MatchCandidate}s to match the index against
+     */
+    @Nonnull
+    default Iterable<MatchCandidate> createMatchCandidates(@Nonnull RecordMetaData metaData, @Nonnull Index index, boolean reverse) {
+        return Collections.emptyList();
+    }
 }
