+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/IndexMaintainerRegistryImpl.java
@@ -24,7 +24,6 @@ import com.apple.foundationdb.annotation.API;
 import com.apple.foundationdb.record.logging.KeyValueLogMessage;
 import com.apple.foundationdb.record.logging.LogMessageKeys;
 import com.apple.foundationdb.record.metadata.Index;
-import com.apple.foundationdb.record.metadata.IndexValidator;
 import com.apple.foundationdb.record.metadata.MetaDataException;
 import com.apple.foundationdb.record.util.ServiceLoaderProvider;
 import org.slf4j.Logger;
@@ -38,7 +37,7 @@ import java.util.Map;
  * A singleton {@link IndexMaintainerRegistry} that finds {@link IndexMaintainerFactory} classes in the classpath.
  */
 @API(API.Status.INTERNAL)
-public class IndexMaintainerRegistryImpl implements IndexMaintainerRegistry {
+public class IndexMaintainerRegistryImpl implements IndexMaintainerFactoryRegistry {
     @Nonnull
     private static final Logger LOGGER = LoggerFactory.getLogger(IndexMaintainerRegistryImpl.class);
     @Nonnull
@@ -75,21 +74,11 @@ public class IndexMaintainerRegistryImpl implements IndexMaintainerRegistry {
 
     @Nonnull
     @Override
-    public IndexValidator getIndexValidator(@Nonnull Index index) {
+    public IndexMaintainerFactory getIndexMaintainerFactory(@Nonnull final Index index) {
         final IndexMaintainerFactory factory = registry.get(index.getType());
         if (factory == null) {
             throw new MetaDataException("Unknown index type for " + index);
         }
-        return factory.getIndexValidator(index);
-    }
-
-    @Nonnull
-    @Override
-    public IndexMaintainer getIndexMaintainer(@Nonnull IndexMaintainerState state) {
-        final IndexMaintainerFactory factory = registry.get(state.index.getType());
-        if (factory == null) {
-            throw new MetaDataException("Unknown index type for " + state.index);
-        }
-        return factory.getIndexMaintainer(state);
+        return factory;
     }
 }
